Fix security

+++ src/main/java/pe/edu/upc/Karwas/service/RoleService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import pe.edu.upc.Karwas.model.entity.Role;
 
 
-public interface RoleService extends CrudService<Role, Integer>{
+public interface RoleService extends CrudService<Role, Long>{
 
 	List<Role> findByNameRole( String nameRole ) throws Exception;
